Deprecate `infinite`
It's no longer documented, and it will be removed completely in a future version.  This feature is no longer supported.

This feature isn't very useful, and it's easily misused and misunderstood by consumers.  Consumers who want an infinite mode can code it easily themselves, and that way they're much more likely to understand it.